requirements.txt
anyio==4.11.0
APScheduler==3.11.0
cachetools==4.2.2
certifi==2025.8.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-dotenv==1.1.1
python-telegram-bot==22.5
pytz==2025.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
telegram==0.0.1
tornado==6.5.2
tzdata==2025.2
tzlocal==5.3.1
